src/services/model/requirements.txt
Flask
psycopg2-binary
python-dotenv
scikit-learn
numpy

